first version of the sorting algorithm
converted most javascript to ecma modules.

Criteria and technologies are loaded as classes ( models/Criterion and models/Technology )
Criteria fire update event when modified

The sorter is its own class. It update the array of sorted technologies when criteria are updated. It fires a sorted event.

test.mjs is a debug/test module before UI is implemented