Implementing the Sorting algorithm  (#2)
The sorter is its own class. It update the array of sorted technologies when criteria are updated. It fires a sorted event.
Criteria and technologies are loaded as classes ( Models/Criterion and Models/Technology )
Criteria fire update event when modified

Converted most javascript to ecma modules to have browser and node support. 

test.mjs is a debug/test module before UI is implemented

Document in markdown in ./docs