Feat: add sorting in CSV
chore(package.json): add @prettier/plugin-pug and prettier as devDependencies
The @prettier/plugin-pug and prettier packages are added as devDependencies in the package.json file. This is done to improve the code formatting and maintain consistency in the project.

fix(Downloader.mjs): remove console.log statement
The console.log statement in the Downloader class is removed. This is done to clean up the code and remove unnecessary logging.

fix(label.mjs): add condition to moveHandle method
The moveHandle method in the Label class now has a condition to check if the ellipse is null before moving the handle. This is done to prevent errors when the moveHandle method is called without a created handle.

fix(capuchin.mjs): refactor loadFromCSV function
The loadFromCSV function in the capuchin.mjs file is refactored to improve readability and maintainability. The function now includes helper functions to parse the criterion name and order, reduce the minimum and maximum values, and filter out unnecessary fields. This improves the code structure and makes it easier to understand.

fix(index.pug): remove unused code
The unused code related to the M.Tabs instance is removed from the index.pug file. This is done to clean up the code and remove unnecessary code.

fix(table.pug): refactor table rendering
The table rendering in the table.pug file is refactored to improve readability and maintainability. The criteria are now sorted based on their weight and only criteria with a weight greater than 0 are rendered. This improves the table structure and makes it easier to understand.